changing project structure to package the code: - moved Replicon class declaration to package __init__ - added repl number cleaning to clean_xl function on replicon_ws - changed the excel macros to call procedures from the new replicon package
