+++ repl_uploader/pyqt/qrcode.ui
